Finished up blueprints (mostly) Changed powerout config for dynamos to a double, to allow for more precise modifiers, fixes #239 fixed derpy in-hand rendering for fences

+++ src/main/java/blusunrize/immersiveengineering/common/blocks/metal/TileEntityDynamo.java
@@ -20,7 +20,7 @@ public class TileEntityDynamo extends TileEntityIEBase implements IEnergyConnect
 	{
 		if(side!=ForgeDirection.OPPOSITES[facing])
 			return;
-		int output = (int) (Config.getInt("dynamo_output") * rotation);
+		int output = (int) (Config.getDouble("dynamo_output") * rotation);
 		for(int i=0; i<6; i++)
 		{
 			ForgeDirection fd = ForgeDirection.VALID_DIRECTIONS[i];
